+++ src/components/Project/ProjectCard.jsx
@@ -1,3 +1,5 @@
+import React from 'react';
+import { Link } from 'react-router-dom';
 import {
   Card,
   CardHeader,
@@ -9,18 +11,26 @@ import {
 } from '@material-tailwind/react';
 
 // this infirmation will be fassed as a prop from
-const project = {
-  name: 'Revolutionary New Social Platform',
-  description: 'project description',
-  technologies: ['React', 'Tailwind', 'Node.js', 'Express'],
-  status: 'Seeking Team Members',
-  rolesNeeded: ['UI designer', 'React Developer', 'Backend Developer'],
-};
+// const project = {
+//   name: 'Revolutionary New Social Platform',
+//   description: 'project description',
+//   technologies: ['React', 'Tailwind', 'Node.js', 'Express'],
+//   status: 'Seeking Team Members',
+//   rolesNeeded: ['UI designer', 'React Developer', 'Backend Developer'],
+// };
 
-const ProjectCard = ({}) => {
+const ProjectCard = ({ project }) => {
   return (
     <Card className="max-w-[13rem] min-w-[12rem] overflow-hidden bg-gray/10 rounded-xl border border-transparent hover:border-blue/30 mb-6 mr-8 p-4">
-      <a href="url-to-the-individual-project">
+     <Link to={`/projects/${project._id}`} 
+        state={{ 
+          projectId: project._id, 
+          projectTitle: project.title, 
+          projectStatus: project.status, 
+          projectDesc: project.description,
+          projectTechnologies: project.technologies,
+          projectRolesNeeded: project.rolesNeeded 
+        }}> 
         <CardHeader
           floated={false}
           shadow={false}
@@ -34,7 +44,7 @@ const ProjectCard = ({}) => {
                 alt="project logo"
             />
             <Typography variant="h3" className="text-xs font-[Jura] text-gray pl-4 font-medium">
-            { project.name}
+            { project.title}
             </Typography>
         </div>
         </CardHeader>
@@ -56,45 +66,36 @@ const ProjectCard = ({}) => {
             <div className="font-normal text-[10px] text-green text-[11px] font-medium">
               Technologies:
             </div>
-            {project.technologies.map((item, index) => (
-              <span key={index} className="font-sans font-extralight text-[10px]">
-                {' '}
-                {item + ', '}{' '}
-              </span>
-            ))}
+            <ul>{renderTechnologies(project.technologies)}</ul>
           </div>
           <Typography className="font-[Jura] text-[10px] text-center text-blue font-medium py-0.5 my-4 border-[0.5px] border-blue/80 rounded-full">
             {project.status}
           </Typography>
           <div variant="lead" color="gray" className="font-normal text-[10px]">
-            <div className="font-normal text-[10px] text-green text-[11px] font-medium">
-              Rooles Needed:
-            </div>
-            {project.rolesNeeded.map((item, index) => (
-              <span key={index} className="font-sans font-extralight text-[10px]">
-                {' '}
-                {item + ', '}{' '}
-              </span>
-            ))}
+            {/* <div className="font-normal text-[10px] text-green text-[11px] font-medium">
+              Roles Needed:
+            </div> */}
           </div>
         </CardBody>
-      </a>
       {/* This will be part of Authorized rout */}
       <CardFooter className="flex items-center justify-start pt-2">
-        <div className="font-normal text-[10px] text-green text-[11px] font-medium pr-4">
+          {/* variant="h3"
+          className="text-[11px] text-right font-[Jura] pl-4 mb-2 mr-2 font-medium" */}
+          See more...
+        {/* <div className="font-normal text-[10px] text-green text-[11px] font-medium pr-4">
           Participants:
-        </div>
-        <div className="flex items-center -space-x-3">
-          <Tooltip content="Natali Craig" className="font-sans font-extralight text-[10px] px-2">
+        </div> */}
+        {/* <div className="flex items-center -space-x-3">
+          <Tooltip content="Natali Craig" className="font-sans font-extralight text-[10px] px-2"> */}
             {/* Avatar can be linked to the profile of the user on Authorized rout */}
-            <a href="/profile" className="hover:z-10">
+            {/* <a href="/profile" className="hover:z-10">
               <Avatar
                 alt="natali craig"
                 src="https://images.unsplash.com/photo-1580489944761-15a19d654956?ixlib=rb-1.2.1&ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&auto=format&fit=crop&w=1061&q=80"
                 className="border-2 border-blue  h-8 w-8 rounded-full hover:border-green hover:cursor-pointer"
               />
-            </a>
-          </Tooltip>
+            </a> */}
+          {/* </Tooltip>
           <Tooltip content="Tania Andrew" className="font-sans font-extralight text-[10px] px-2">
             <a href="/profile" className="hover:z-10">
               <Avatar
@@ -106,9 +107,32 @@ const ProjectCard = ({}) => {
               />
             </a>
           </Tooltip>
-        </div>
+        </div> */}
       </CardFooter>
-    </Card>
+      </Link>
+  </Card>
+
+  );
+};
+
+const renderTechnologies = technologies => {
+  if (!technologies) return null;
+
+  const allTech = [];
+  for (const type in technologies) {
+    allTech.push(...technologies[type]);
+  }
+
+  const displayedTech = allTech.slice(0, 4);
+  const etcTechCount = allTech.length - displayedTech.length;
+
+  return (
+    <div>
+      {displayedTech.map((tech, index) => (
+        <li key={index}>• {tech}</li>
+      ))}
+      {etcTechCount > 0 && <li>etc. ({etcTechCount} more)</li>}
+    </div>
   );
 };
